fix MathieuRA comments
