--- @xen-orchestra/rest-api/src/backup-repositories/backup-repositories.service.mts
@@ -1,15 +1,0 @@
-import { IdOr, XoBackupRepository } from '@vates/types'
-
-export class BackupRepositoriesService {
-  public isRepositoryInRemotes(
-    remotes: IdOr<XoBackupRepository['id']> | undefined,
-    repositoryId: XoBackupRepository['id']
-  ): boolean {
-    if (remotes === undefined) {
-      return false
-    }
-    const { id } = remotes
-    const ids = typeof id === 'string' ? [id] : id.__or
-    return ids.includes(repositoryId)
-  }
-}
